latest update to basic report structure
It has table, outline for workspace and project, but am currently trying to figure out how to make them go in the correct order.  Still need to make projects and tasks under its own workspace... right now it is very rigid and basic structure with no connection to the data.